Move metadata table to metadata feature
