+++ idea/frontend/src/features/metadata/index.ts
@@ -1,3 +1,4 @@
 import { useMetadata } from './hooks';
+import { MetadataTable } from './ui';
 
-export { useMetadata };
+export { useMetadata, MetadataTable };
