+++ idea/frontend/src/features/metadata/ui/index.ts
@@ -1,0 +1,3 @@
+import { MetadataTable } from './metadata-table';
+
+export { MetadataTable };
