+++ idea/frontend/src/features/metadata/ui/metadata-table/index.ts
@@ -1,0 +1,3 @@
+import { MetadataTable } from './metadata-table';
+
+export { MetadataTable };
