+++ idea/frontend/src/pages/code/ui/Code.tsx
@@ -13,11 +13,10 @@ import { UILink } from 'shared/ui/uiLink';
 import { Table, TableRow } from 'shared/ui/table';
 import { IdBlock } from 'shared/ui/idBlock';
 import { ProgramsList } from 'pages/programs/ui/programsList';
-import { MetadataDetails } from 'pages/program/ui/metadataDetails';
+import { MetadataTable, useMetadata } from 'features/metadata';
 import { ReactComponent as PlusSVG } from 'shared/assets/images/actions/plus.svg';
 import { ReactComponent as AddMetaSVG } from 'shared/assets/images/actions/addMeta.svg';
 import { ICode } from 'entities/code';
-import { useMetadata } from 'features/metadata';
 
 import styles from './Code.module.scss';
 
@@ -81,7 +80,7 @@ const Code = () => {
 
           <div>
             <h2 className={styles.heading}>Metadata</h2>
-            <MetadataDetails metadata={metadata} isLoading={!isMetadataReady} />
+            <MetadataTable metadata={metadata} isLoading={!isMetadataReady} />
           </div>
         </div>
