+++ idea/frontend/src/pages/program/program.module.scss
@@ -1,7 +1,7 @@
-@use 'shared/assets/styles/_mixins.scss' as *;
-@use 'shared/assets/styles/_variables.scss' as *;
-@use 'shared/assets/styles/_animations.scss' as *;
-@use 'shared/assets/styles/_shared.scss' as *;
+@use 'shared/assets/styles/mixins' as *;
+@use 'shared/assets/styles/variables' as *;
+@use 'shared/assets/styles/animations' as *;
+@use 'shared/assets/styles/shared' as *;
 
 .header {
   display: flex;
