+++ idea/frontend/src/pages/program/program.tsx
@@ -5,7 +5,7 @@ import { useAccount } from '@gear-js/react-hooks';
 import { generatePath, useParams } from 'react-router-dom';
 
 import { useMetadataUpload, useModal, useProgram } from 'hooks';
-import { ProgramStatus , ProgramTable } from 'features/program';
+import { ProgramStatus, ProgramTable } from 'features/program';
 import { ProgramMessages } from 'widgets/programMessages';
 import { PathParams } from 'shared/types';
 import { getShortName } from 'shared/helpers';
@@ -15,12 +15,10 @@ import { UILink } from 'shared/ui/uiLink';
 import { ReactComponent as SendSVG } from 'shared/assets/images/actions/send.svg';
 import { ReactComponent as ReadSVG } from 'shared/assets/images/actions/read.svg';
 import { ReactComponent as AddMetaSVG } from 'shared/assets/images/actions/addMeta.svg';
-import { useMetadata } from 'features/metadata';
+import { useMetadata, MetadataTable } from 'features/metadata';
 import { VoucherTable } from 'features/voucher';
 
-
-import { MetadataDetails } from './metadataDetails';
-import styles from './Program.module.scss';
+import styles from './program.module.scss';
 
 const Program = () => {
   const { account } = useAccount();
@@ -95,7 +93,7 @@ const Program = () => {
 
           <div>
             <Subheader title="Metadata" />
-            <MetadataDetails metadata={metadata} isLoading={!isMetadataReady} />
+            <MetadataTable metadata={metadata} isLoading={!isMetadataReady} />
           </div>
         </div>
