--- idea/frontend/src/pages/program/ui/metadataDetails/index.ts
@@ -1,3 +1,0 @@
-import { MetadataDetails } from './MetadataDetails';
-
-export { MetadataDetails };
